Phase 2 step 7: Extract rates/fuel/metal fetchers ONLY (keep OSINT in dashboard_tui)
